+++ MusicMachine.vcxproj
@@ -134,13 +134,13 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="include\JSON\JSON.cpp" />
-    <ClCompile Include="src\App\NodeNetworkRenderer.cpp" />
+    <ClCompile Include="src\App\JSON.cpp" />
+    <ClCompile Include="src\App\Nodes\NodeNetworkRenderer.cpp" />
     <ClCompile Include="src\Engine\DrawList.cpp" />
-    <ClCompile Include="src\App\NodeTypes.cpp" />
-    <ClCompile Include="src\App\Node.cpp" />
-    <ClCompile Include="src\App\NodeNetwork.cpp" />
-    <ClCompile Include="src\App\Canvas.cpp" />
+    <ClCompile Include="src\App\Nodes\NodeTypes.cpp" />
+    <ClCompile Include="src\App\Nodes\Node.cpp" />
+    <ClCompile Include="src\App\Nodes\NodeNetwork.cpp" />
+    <ClCompile Include="src\App\Nodes\Canvas.cpp" />
     <ClCompile Include="src\App\App.cpp" />
     <ClCompile Include="src\Engine\Console.cpp" />
     <ClCompile Include="src\Engine\Engine.cpp" />
@@ -172,10 +172,10 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="include\App\App.h" />
-    <ClInclude Include="include\App\Canvas.h" />
-    <ClInclude Include="include\App\Node.h" />
-    <ClInclude Include="include\App\NodeNetwork.h" />
-    <ClInclude Include="include\App\NodeTypes.h" />
+    <ClInclude Include="include\App\Nodes\Canvas.h" />
+    <ClInclude Include="include\App\Nodes\Node.h" />
+    <ClInclude Include="include\App\Nodes\NodeNetwork.h" />
+    <ClInclude Include="include\App\Nodes\NodeTypes.h" />
     <ClInclude Include="include\BBox.h" />
     <ClInclude Include="include\Engine.h" />
     <ClInclude Include="include\Engine\Console.h" />
@@ -195,7 +195,7 @@
     <ClInclude Include="include\Engine\Texture\RenderTexture.h" />
     <ClInclude Include="include\Engine\Texture\Texture.h" />
     <ClInclude Include="include\Engine\Texture\Texture2D.h" />
-    <ClInclude Include="include\JSON\JSON.h" />
+    <ClInclude Include="include\App\JSON.h" />
     <ClInclude Include="include\Matrix.h" />
     <ClInclude Include="include\Quaternion.h" />
     <ClInclude Include="include\Random.h" />
