+++ include/App/Nodes/Node.h
@@ -2,7 +2,7 @@
 #include "Vector.h"
 #include "BBox.h"
 #include <vector>
-#include "JSON/JSON.h"
+#include "App/JSON.h"
 
 struct Node;
